+++ lib/beacon/application.ex
@@ -5,6 +5,8 @@ defmodule Beacon.Application do
 
   @impl true
   def start(_type, _args) do
+    # Force loading ErrorHandler before it's needed,
+    # otherwise we get race conditions auto loading Beacon resource modules.
     {:module, _} = Code.ensure_loaded(Beacon.ErrorHandler)
 
     # Starts just the minimum required apps for beacon to work.
@@ -18,6 +20,6 @@ defmodule Beacon.Application do
     # TODO: scope by site
     :ets.new(:beacon_assets, [:set, :named_table, :public, read_concurrency: true])
 
-    Supervisor.start_link(children, strategy: :one_for_one, name: __MODULE__)
+    Supervisor.start_link(children, strategy: :one_for_one, name: Beacon.Supervisor)
   end
 end