+++ lib/beacon/boot.ex
@@ -15,27 +15,16 @@ defmodule Beacon.Boot do
     Beacon.Registry.via({site, __MODULE__})
   end
 
-  def init(%{site: site, mode: :manual}) when is_atom(site) do
-    Logger.debug("Beacon.Boot is disabled for site #{site} on manual mode")
+  def init(%{site: site, mode: mode}) when is_atom(site) and mode in [:manual, :testing] do
+    Logger.debug("Beacon.Boot is disabled for site #{site} on #{mode} mode")
     :ignore
   end
 
-  def init(%{site: site, mode: :testing}) when is_atom(site) do
-    Logger.debug("Beacon.Boot is disabled for site #{site} on testing mode")
-
-    # reload modules that are expected to be available, even empty
-    Beacon.Loader.reload_routes_module(site)
-    Beacon.Loader.reload_components_module(site)
-    Beacon.Loader.reload_live_data_module(site)
-
-    :ignore
-  end
-
-  def init(%{site: site, mode: :live}) when is_atom(site) do
+  def init(%{site: site, mode: :live} = config) when is_atom(site) do
     Logger.info("Beacon.Boot booting site #{site}")
     task_supervisor = Beacon.Registry.via({site, TaskSupervisor})
 
-    # temporary disable module reloading so we can populate data more efficiently
+    # temporary disable module loading so we can populate data more efficiently
     %{mode: :manual} = Beacon.Config.update_value(site, :mode, :manual)
     Beacon.Loader.populate_default_media(site)
     Beacon.Loader.populate_default_components(site)
@@ -45,21 +34,40 @@ defmodule Beacon.Boot do
 
     %{mode: :live} = Beacon.Config.update_value(site, :mode, :live)
 
-    # still needed to test Beacon itself
-    Beacon.Loader.reload_routes_module(site)
-    Beacon.Loader.reload_components_module(site)
-
-    assets = [
-      Task.Supervisor.async(task_supervisor, fn -> Beacon.Loader.reload_runtime_js(site) end),
-      Task.Supervisor.async(task_supervisor, fn -> Beacon.Loader.reload_runtime_css(site) end)
+    tasks = [
+      Task.Supervisor.async(task_supervisor, fn -> Beacon.Loader.load_runtime_js(site) end),
+      Task.Supervisor.async(task_supervisor, fn -> Beacon.Loader.load_runtime_css(site) end)
+      | warm_pages_async(task_supervisor, config)
     ]
 
     # TODO: revisit this timeout after we upgrade to Tailwind v4
-    Task.await_many(assets, :timer.minutes(5))
+    Task.await_many(tasks, :timer.minutes(5))
 
     # TODO: add telemetry to measure booting time
     Logger.info("Beacon.Boot finished booting site #{site}")
 
     :ignore
+  end
+
+  defp warm_pages_async(task_supervisor, config) do
+    pages =
+      case config.page_warming do
+        {:shortest_paths, count} ->
+          Logger.info("Beacon.Boot warming pages - #{count} shortest paths")
+          Beacon.Content.list_published_pages(config.site, sort: {:length, :path}, limit: count)
+
+        {:specify_paths, paths} ->
+          Logger.info("Beacon.Boot warming pages - specified paths")
+          Beacon.Content.list_published_pages_for_paths(config.site, paths)
+
+        :none ->
+          Logger.info("Beacon.Boot page warming disabled")
+          []
+      end
+
+    Enum.map(pages, fn page ->
+      Logger.info("Beacon.Boot warming page #{page.id} #{page.path}")
+      Task.Supervisor.async(task_supervisor, fn -> Beacon.Loader.load_page_module(config.site, page.id) end)
+    end)
   end
 end