+++ lib/beacon/web/beacon_assigns.ex
@@ -42,7 +42,8 @@ defmodule Beacon.Web.BeaconAssigns do
               info_handlers_module: nil,
               event_handlers_module: nil,
               live_data_keys: [],
-              live_path: []
+              live_path: [],
+              variant_roll: nil
             }
 
   @doc false
@@ -52,20 +53,28 @@ defmodule Beacon.Web.BeaconAssigns do
   end
 
   @doc false
-  def new(site, %Beacon.Content.Page{} = page) do
+  def new(site, %Beacon.Content.Page{} = page, variant_roll) do
     components_module = Beacon.Loader.Components.module_name(site)
     page_module = Beacon.Loader.Page.module_name(site, page.id)
-    %__MODULE__{site: site, private: %{components_module: components_module, page_module: page_module}}
+
+    %__MODULE__{
+      site: site,
+      private: %{
+        components_module: components_module,
+        page_module: page_module,
+        variant_roll: variant_roll
+      }
+    }
   end
 
   @doc false
-  def new(site, %Beacon.Content.Page{} = page, live_data, path_info, query_params)
-      when is_atom(site) and is_map(live_data) and is_list(path_info) and is_map(query_params) do
+  def new(site, %Beacon.Content.Page{} = page, live_data, path_info, query_params, source, variant_roll \\ nil)
+      when is_atom(site) and is_map(live_data) and is_list(path_info) and is_map(query_params) and source in [:beacon, :admin] do
     %{site: ^site} = page
     page_module = Beacon.Loader.Page.module_name(site, page.id)
     live_data = Beacon.Web.DataSource.live_data(site, path_info, Map.drop(query_params, ["path"]))
     path_params = Beacon.Router.path_params(page.path, path_info)
-    page_title = Beacon.Web.DataSource.page_title(site, page.id, live_data)
+    page_title = Beacon.Web.DataSource.page_title(site, page.id, live_data, source)
     components_module = Beacon.Loader.Components.module_name(site)
     info_handlers_module = Beacon.Loader.InfoHandlers.module_name(site)
     event_handlers_module = Beacon.Loader.EventHandlers.module_name(site)
@@ -81,7 +90,8 @@ defmodule Beacon.Web.BeaconAssigns do
         info_handlers_module: info_handlers_module,
         event_handlers_module: event_handlers_module,
         live_data_keys: Map.keys(live_data),
-        live_path: path_info
+        live_path: path_info,
+        variant_roll: variant_roll
       }
     }
   end