+++ lib/beacon/web/live/page_live.ex
@@ -19,23 +19,39 @@ defmodule Beacon.Web.PageLive do
     %{"path" => path} = params
     %{"beacon_site" => site} = session
 
-    # Use Beacon custom error handler to automatically load modules on-demand
-    if Beacon.Config.fetch!(site).mode == :live do
-      Beacon.ErrorHandler.enable(site)
-      if connected?(socket), do: :ok = Beacon.PubSub.subscribe_to_page(site, path)
+    if Beacon.Config.fetch!(site).mode == :live and connected?(socket) do
+      :ok = Beacon.PubSub.subscribe_to_page(site, path)
     end
 
+    variant_roll =
+      case session["beacon_variant_roll"] do
+        nil ->
+          Logger.warning("""
+          Beacon.Plug is missing from the Router pipeline.
+
+          Page Variants will not be used.
+          """)
+
+          nil
+
+        roll ->
+          roll
+      end
+
     page = RouterServer.lookup_page!(site, path)
-    socket = Component.assign(socket, beacon: BeaconAssigns.new(site, page))
+    socket = Component.assign(socket, beacon: BeaconAssigns.new(site, page, variant_roll))
 
     {:ok, socket, layout: {Beacon.Web.Layouts, :dynamic}}
   end
 
   def render(assigns) do
-    %{beacon: %{private: %{page_module: page_module}}} = assigns
+    %{beacon: %{site: site, private: %{page_module: page_module}}} = assigns
 
-    page_module
-    |> Beacon.apply_mfa(:page, [])
+    # do not allow overwritring site to avoid rendering a leaked page
+    %{site: ^site} = Beacon.apply_mfa(site, page_module, :page_assigns, [[:site]])
+
+    site
+    |> Beacon.apply_mfa(page_module, :page, [])
     |> Lifecycle.Template.render_template(assigns, __ENV__)
   end
 
@@ -55,11 +71,12 @@ defmodule Beacon.Web.PageLive do
   end
 
   def handle_info(msg, socket) do
-    %{page_module: page_module, live_path: live_path, info_handlers_module: info_handlers_module} = socket.assigns.beacon.private
-    %{site: site, id: page_id} = Beacon.apply_mfa(page_module, :page_assigns, [[:site, :id]])
+    %{beacon: %{site: site, private: %{page_module: page_module, live_path: live_path, info_handlers_module: info_handlers_module}}} = socket.assigns
+    %{site: ^site, id: page_id} = Beacon.apply_mfa(site, page_module, :page_assigns, [[:site, :id]])
 
     result =
       Beacon.apply_mfa(
+        site,
         info_handlers_module,
         :handle_info,
         [msg, socket],
@@ -77,11 +94,14 @@ defmodule Beacon.Web.PageLive do
   end
 
   def handle_event(event_name, event_params, socket) do
-    %{page_module: page_module, live_path: live_path, event_handlers_module: event_handlers_module} = socket.assigns.beacon.private
-    %{site: site, id: page_id} = Beacon.apply_mfa(page_module, :page_assigns, [[:site, :id]])
+    %{beacon: %{site: site, private: %{page_module: page_module, live_path: live_path, event_handlers_module: event_handlers_module}}} =
+      socket.assigns
+
+    %{site: ^site, id: page_id} = Beacon.apply_mfa(site, page_module, :page_assigns, [[:site, :id]])
 
     result =
       Beacon.apply_mfa(
+        site,
         event_handlers_module,
         :handle_event,
         [event_name, event_params, socket],
@@ -109,15 +129,17 @@ defmodule Beacon.Web.PageLive do
 
       site ->
         %{"path" => path_info} = params
+
+        if socket.assigns.beacon.site != site do
+          if Beacon.Config.fetch!(site).mode == :live do
+            Beacon.PubSub.unsubscribe_to_page(socket.assigns.beacon.site, path_info)
+            Beacon.PubSub.subscribe_to_page(site, path_info)
+          end
+        end
+
         page = RouterServer.lookup_page!(site, path_info)
         live_data = Beacon.Web.DataSource.live_data(site, path_info, Map.drop(params, ["path"]))
-        beacon_assigns = BeaconAssigns.new(site, page, live_data, path_info, params)
-
-        if socket.assigns.beacon.site != site && Beacon.Config.fetch!(site).mode == :live do
-          Process.put(:__beacon_site__, site)
-          Beacon.PubSub.unsubscribe_to_page(socket.assigns.beacon.site, path_info)
-          Beacon.PubSub.subscribe_to_page(site, path_info)
-        end
+        beacon_assigns = BeaconAssigns.new(site, page, live_data, path_info, params, :beacon, socket.assigns.beacon.private.variant_roll)
 
         socket =
           socket
@@ -129,7 +151,7 @@ defmodule Beacon.Web.PageLive do
           # TODO: remove deprecated @beacon_query_params
           |> Component.assign(:beacon_query_params, beacon_assigns.query_params)
           |> Component.assign(:beacon, beacon_assigns)
-          |> Component.assign(:page_title, Beacon.Web.DataSource.page_title(site, page.id, live_data))
+          |> Component.assign(:page_title, Beacon.Web.DataSource.page_title(site, page.id, live_data, :beacon))
 
         {:noreply, push_event(socket, "beacon:page-updated", %{meta_tags: Beacon.Web.DataSource.meta_tags(socket.assigns)})}
     end