+++ test/beacon/registry_test.exs
@@ -8,6 +8,7 @@ defmodule Beacon.RegistryTest do
              :booted,
              :data_source_test,
              :default_meta_tags_test,
+             :host_test,
              :lifecycle_test,
              :lifecycle_test_fail,
              :my_site,