+++ test/beacon_web/data_source_test.exs
@@ -3,21 +3,33 @@ defmodule Beacon.Web.DataSourceTest do
   use Beacon.Test, site: :my_site
   alias Beacon.Web.DataSource
 
-  test "live_data" do
+  setup do
+    site = default_site()
+
+    # we aren't passing through PageLive in these tests so we have to manually
+    # enable the ErrorHandler and set the site in the Process dictionary
+    # (which would normally happen in the LiveView mount)
+    Process.put(:__beacon_site__, site)
+    Process.flag(:error_handler, Beacon.ErrorHandler)
+
+    [site: site]
+  end
+
+  test "live_data", %{site: site} do
     live_data = beacon_live_data_fixture()
     beacon_live_data_assign_fixture(live_data: live_data, format: :text, key: "name", value: "beacon")
-    assert DataSource.live_data(default_site(), ["foo", "bar"], %{}) == %{name: "beacon"}
+    assert DataSource.live_data(site, ["foo", "bar"], %{}) == %{name: "beacon"}
   end
 
   describe "page_title" do
-    test "renders static content" do
-      page = beacon_published_page_fixture(site: default_site(), title: "my title")
-      assert DataSource.page_title(page.site, page.id, %{}) == "my title"
+    test "renders static content", %{site: site} do
+      page = beacon_published_page_fixture(site: site, title: "my title")
+      assert DataSource.page_title(page.site, page.id, %{}, :beacon) == "my title"
     end
 
-    test "renders snippet" do
-      page = beacon_published_page_fixture(site: default_site(), title: "{{ page.path | upcase }}")
-      assert DataSource.page_title(page.site, page.id, %{}) == "/HOME"
+    test "renders snippet", %{site: site} do
+      page = beacon_published_page_fixture(site: site, title: "{{ page.path | upcase }}")
+      assert DataSource.page_title(page.site, page.id, %{}, :beacon) == "/HOME"
     end
   end
 end