+++ test/beacon_web/live/page_live_test.exs
@@ -68,7 +68,7 @@ defmodule Beacon.Web.Live.PageLiveTest do
           @beacon.query_params=<%= @beacon.query_params["query"] %>
 
           <.page_link path="/about">go_to_about_page</.page_link>
-          <.link patch={"#{Beacon.BeaconTest.Endpoint.url()}/other"}>go_to_other_site</.link>
+          <.link navigate="/other">go_to_other_site</.link>
 
           <.form :let={f} for={%{}} as={:greeting} phx-submit="hello">
             Name: <%= text_input f, :name %>
@@ -171,9 +171,11 @@ defmodule Beacon.Web.Live.PageLiveTest do
     test "patch to another site resets site data", %{conn: conn} do
       {:ok, view, _html} = live(conn, "/home/hello")
 
-      view
-      |> element("a", "go_to_other_site")
-      |> render_click()
+      {:ok, view, _html} =
+        view
+        |> element("a", "go_to_other_site")
+        |> render_click()
+        |> follow_redirect(conn, "/other")
 
       assert has_element?(view, "h1", "not_booted")
     end
@@ -276,7 +278,7 @@ defmodule Beacon.Web.Live.PageLiveTest do
     test "update resource links on layout publish", %{conn: conn, layout: layout} do
       {:ok, layout} = Content.update_layout(layout, %{"resource_links" => [%{"rel" => "stylesheet", "href" => "color.css"}]})
       {:ok, layout} = Content.publish_layout(layout)
-      Beacon.Loader.reload_layout_module(layout.site, layout.id)
+      Beacon.Loader.load_layout_module(layout.site, layout.id)
       {:ok, _view, html} = live(conn, "/home/hello")
       assert html =~ ~S|<link href="color.css" rel="stylesheet"/>|
     end
@@ -384,7 +386,7 @@ defmodule Beacon.Web.Live.PageLiveTest do
       assert html =~ "component_test_v1"
 
       Content.update_component(component, %{template: "component_test_v2"})
-      Beacon.Loader.reload_components_module(component.site)
+      Beacon.Loader.load_components_module(component.site)
 
       {:ok, _view, html} = live(conn, "/component_test")
       assert html =~ "component_test_v2"